requirements.txt
datasets==2.13.0
omegaconf==2.3.0
peft==0.4.0
PyYAML==6.0.1
PyYAML==6.0.1
torch==2.0.1+cu118
transformers==4.33.2
trl==0.7.2
